@@ -1,24 +1,26 @@
 module github.com/takutakahashi/any-exporter
 
-go 1.13
+go 1.20
 
 require (
-	github.com/containerd/containerd v1.3.3
-	github.com/containerd/ttrpc v1.0.0 // indirect
-	github.com/containerd/typeurl v1.0.0 // indirect
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
-	github.com/docker/go-events v0.0.0-20190806004212-e31b211e4f1c // indirect
-	github.com/gogo/protobuf v1.3.1 // indirect
-	github.com/google/martian v2.1.0+incompatible
-	github.com/imdario/mergo v0.3.9 // indirect
 	github.com/labstack/echo v3.3.10+incompatible
-	github.com/labstack/gommon v0.3.0 // indirect
-	github.com/opencontainers/go-digest v1.0.0-rc1 // indirect
-	github.com/opencontainers/image-spec v1.0.1 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/sirupsen/logrus v1.5.0
-	github.com/urfave/cli v1.22.3
-	google.golang.org/grpc v1.28.0 // indirect
+	github.com/sirupsen/logrus v1.9.3
+	github.com/urfave/cli v1.22.14
+	golang.org/x/sync v0.6.0
 	gopkg.in/yaml.v1 v1.0.0-20140924161607-9f9df34309c0
-	gotest.tools v2.2.0+incompatible // indirect
+)
+
+require (
+	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
+	github.com/labstack/gommon v0.4.2 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/russross/blackfriday/v2 v2.1.0 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasttemplate v1.2.2 // indirect
+	golang.org/x/crypto v0.20.0 // indirect
+	golang.org/x/net v0.21.0 // indirect
+	golang.org/x/sys v0.17.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
 )
