@@ -1,7 +1,7 @@
 package config
 
 import (
-	"io/ioutil"
+	"os"
 
 	"gopkg.in/yaml.v1"
 )
@@ -16,11 +16,12 @@ type Config struct {
 type MetricsConfig struct {
 	WorkDir    string   `yaml:"workdir"`
 	ScriptPath string   `yaml:"scriptPath"`
+	Resolution string   `yaml:"resolution"`
 	Args       []string `yaml:"args"`
 }
 
 func Load(configPath string) (Config, error) {
-	str, err := ioutil.ReadFile(configPath)
+	str, err := os.ReadFile(configPath)
 	if err != nil {
 		return Config{}, err
 	}
@@ -29,5 +30,23 @@ func Load(configPath string) (Config, error) {
 	if err != nil {
 		return Config{}, err
 	}
-	return config, nil
+	return FillDefaults(config), nil
+}
+
+func FillDefaults(config Config) Config {
+	if config.Host == "" {
+		config.Host = "localhost"
+	}
+	if config.Port == "" {
+		config.Port = "8080"
+	}
+	if config.Resolution == "" {
+		config.Resolution = "1m"
+	}
+	for i, metric := range config.Metrics {
+		if metric.Resolution == "" {
+			config.Metrics[i].Resolution = config.Resolution
+		}
+	}
+	return config
 }
